# feat(rust/catalyst-voting): Parallel iter usage for `fn encrypt_vote`