--- a/rust/catalyst-voting/Cargo.toml
+++ b/rust/catalyst-voting/Cargo.toml
@@ -23,6 +23,7 @@ rand_chacha = "0.3.1"
 curve25519-dalek = { version = "4.1.3", features = ["digest", "rand_core"] }
 ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 blake2b_simd = "1.0.2"
+rayon = "1.10.0"
 
 [dev-dependencies]
 criterion = "0.5.1"