# Description

- Added `rayon` lib
- Updated `encrypt_vote` function with the parallel iterator usage.

Performance measurements on `Apple M1`
Old version:
```
vote protocol benchmark/vote encryption
                        time:   [243.70 µs 243.88 µs 244.14 µs]
```
New version:
```
vote protocol benchmark/vote encryption
                        time:   [139.87 µs 141.72 µs 144.27 µs]
                        change: [-42.351% -41.334% -40.375%] (p = 0.00 < 0.05)
                        Performance has improved.
```